Scope VectorCypher recall-filter tests to the counter-firing gap
Reduce the VectorCypher recall-filter test surface to the one thing not already
covered elsewhere: that the two service-level filter counters fire on their
declared conditions.

- Drop tests/recall/test_filter_vectorcypher.py: its row-set parity duplicates
  the conformance postgres leg, its pushdown / partial-failure spies duplicate
  the pushdown-spy suite, and its live scenarios ran in no CI job (tests/recall/
  is collected only by the no-DB test-unit job, so they skipped silently).
- Add tests/unit/engines/vectorcypher/test_filter_counters.py: hermetic spies
  that graph_channel_empty and under_filled fire on their declared conditions,
  each paired with a silent-control. Runs in test-unit.
- Add tests/integration/test_vectorcypher_filter_counters.py: real PG+Neo4j proof
  that graph_channel_empty fires when a metadata filter empties a genuine,
  non-empty graph channel (pre-filter vacuity guard). Carries integration +
  filter_enforcement markers so the test-integration job runs it against both DBs.
- Register the filter_enforcement marker.
